update destination driver

+++ app/models/destination_driver.py
@@ -4,8 +4,8 @@ class Destination_Driver(db.Model):
     __tablename__ = 'destination_driver'
     id = db.Column(db.Integer, primary_key=True)
     driver_id = db.Column(db.Integer, db.ForeignKey(Users.id))
-    destination_lat = db.Column(db.Double, nullable=False)
-    destination_lng = db.Column(db.Double, nullable=False)
-    origin_lat = db.Column(db.Double, nullable=False)
-    origin_lng = db.Column(db.Double, nullable=False)
+    destination_lat = db.Column(db.Float, nullable=False)
+    destination_lng = db.Column(db.Float, nullable=False)
+    origin_lat = db.Column(db.Float, nullable=False)
+    origin_lng = db.Column(db.Float, nullable=False)
